+++ src/proto_alpha/lib_delegate/state_transitions.ml
@@ -466,7 +466,7 @@ let repropose_block_action state delegate round (proposal : proposal) =
           (* Add the proposal and pqc consensus operations to the
              mempool *)
           List.fold_left
-            (fun set op -> Operation_pool.OpSet.add op set)
+            (fun set op -> Operation_pool.Operation_set.add op set)
             mempool_consensus_operations
             (List.map Operation.pack proposal.block.quorum
             @ List.map Operation.pack prequorum.preendorsements)
@@ -491,7 +491,7 @@ let repropose_block_action state delegate round (proposal : proposal) =
             ~endorsement_filter
             ~preendorsement_filter
             all_consensus_operations
-          |> OpSet.elements)
+          |> Operation_set.elements)
       in
       let payload_hash = proposal.block.payload_hash in
       let payload_round = proposal.block.payload_round in